==> Auto Insurance Claims/Notebooks/00_data.ipynb <==
import os
import zipfile
from kaggle.api.kaggle_api_extended import KaggleApi

# Set up credentials (adjust the path as needed)
os.environ['KAGGLE_CONFIG_DIR'] = "./.kaggle"

# Initialize Kaggle API
api = KaggleApi()
api.authenticate()

# Download the dataset
api.dataset_download_files('buntyshah/auto-insurance-claims-data', path='../data', unzip=True)
