==> Auto Insurance Claims/Notebooks/02_Feature_Engineering_SD.ipynb <==
# Feature Engineering on Structured Data

# Import dataset
import pandas as pd 
import numpy as np

X_structured = pd.read_csv('../data/X_structured.csv')
y = pd.read_csv('../data/y.csv')

from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline

num_features = X_structured.select_dtypes(include=np.number).columns.tolist()
cat_features = X_structured.select_dtypes(include='object').columns.tolist()


preprocessor = ColumnTransformer(transformers=[
    ('num', StandardScaler(), num_features),
    ('cat', OneHotEncoder(handle_unknown='ignore'), cat_features)
])

# save preprocessor to disk
import joblib   
joblib.dump(preprocessor, '../data/preprocessor.pkl')
