==> Auto Insurance Claims/Notebooks/04_XGBoost_SD.ipynb <==
# XGBoost on Structured Data

from xgboost import XGBClassifier
from sklearn.pipeline import make_pipeline
import pandas as pd

X_structured = pd.read_csv('../data/X_structured.csv')
y = pd.read_csv('../data/y.csv')

# import the preprocessor
import joblib
preprocessor = joblib.load('../data/preprocessor.pkl')

xgb_model = make_pipeline(preprocessor, XGBClassifier(n_estimators=100, use_label_encoder=False, eval_metric='logloss'))
xgb_model.fit(X_structured, y)
