==> Auto Insurance Claims/Notebooks/05_Meta_Model.ipynb <==
# Meta Model (Ensemble)

import pandas as pd
import numpy as np

# Import xgb.model
import joblib
xgb_model = joblib.load('../data/xgb_model.pkl')

predictions = np.load("../data/bert_predictions.npy")
# Create a DataFrame from the predictions
predictions_df = pd.DataFrame(predictions, columns=['prediction'])
# Save the predictions DataFrame to a CSV file
predictions_df.to_csv('../data/bert_predictions.csv', index=False)

# load X_structured data
X_structured = pd.read_csv('../data/X_structured.csv')

# load y data
y = pd.read_csv('../data/y.csv')

# Get predictions from both models
bert_preds = predictions_df['prediction'].values
xgb_preds = xgb_model.predict_proba(X_structured)[:, 1]

# Stack and create meta features
ensemble_X = np.vstack((bert_preds, xgb_preds)).T

# Save the ensemble features
np.save('../data/ensemble_features.npy', ensemble_X)    

from sklearn.linear_model import LogisticRegression
meta_model = LogisticRegression()
meta_model.fit(ensemble_X, y)

# save the meta model
joblib.dump(meta_model, '../data/meta_model.pkl')
