==> Auto Insurance Claims/Notebooks/06_Evaluation.ipynb <==
# Meta Model Evaluation   

import numpy as np
# Load ensemble features
ensemble_X = np.load('../data/ensemble_features.npy')

#load y data
import pandas as pd
y = pd.read_csv('../data/y.csv')['fraud_reported'].values

# Load the meta_model
import joblib
meta_model = joblib.load('../data/meta_model.pkl')

from sklearn.metrics import roc_auc_score, classification_report

meta_preds = meta_model.predict_proba(ensemble_X)[:, 1]
print("AUC:", roc_auc_score(y, meta_preds))
print(classification_report(y, meta_preds > 0.5))
