==> Auto Insurance Claims/Notebooks/08_SHAP.ipynb <==
# SHAP Explainability

# Import xgb.model
import joblib
xgb_model = joblib.load('../data/xgb_model.pkl')

#import X structured data
import pandas as pd 
X_structured = pd.read_csv('../data/X_structured.csv')

import shap

explainer = shap.Explainer(xgb_model.named_steps['xgbclassifier'])
shap_values = explainer(X_structured)
shap.plots.beeswarm(shap_values)
